Minor text adjustments, moved convert_to_tf_coords up to run_detector.py, to avoid importing TF-specific modules.
