+++ api/synchronous/api_core/Dockerfile
@@ -2,7 +2,7 @@
 ARG BASE_IMAGE=tensorflow/tensorflow:1.14.0-gpu-py3
 FROM $BASE_IMAGE
 
-# install and expose redis default port
+# Install and expose redis default port
 RUN  apt-get update && apt-get install -y redis-server
 EXPOSE  6379
 
@@ -28,7 +28,6 @@ RUN apt-get update --fix-missing \
         ffmpeg \
         libsm6 \
         libxext6
-# the base image uses Python 3.6. TODO use TF2 for inference
 
 # Install additional packages
 RUN pip3 install --upgrade pip
@@ -45,7 +44,7 @@ RUN pip3 install \
     jsonpickle
 
 
-# install yolo5 requirements
+# Install yolo5 requirements
 COPY ./yolov5/requirements.txt /yolov5_requirements.txt
 RUN pip3 install -r ./yolov5_requirements.txt
 
@@ -56,7 +55,7 @@ RUN pip3 install -r ./requirements.txt
 COPY ./animal_detection_api /app/animal_detection_api/
 RUN true
 
-#Copy yolo5 code
+# Copy yolo5 code
 COPY ./yolov5 /app/yolov5/
 RUN true
 
@@ -73,8 +72,8 @@ ENV PYTHONUNBUFFERED=TRUE
 
 # Expose the port that is to be used when calling your API
 
-#!!! IF YOU CHANGE THE PORT HERE REMEMBER TO UPDATE supervisord.conf FILE AS WELL
+# If you change the port here, remember to update supervisord.conf 
 EXPOSE 1213
 
-# starts the supervisor.conf (which starts up the detector, flask app and redis server)
+# Starts up the detector, Flask app and Redis server
 ENTRYPOINT [ "/startup.sh" ]
