+++ api/synchronous/api_core/startup.sh
