+++ detection/tf_detector.py
@@ -64,21 +64,6 @@ class TFDetector:
         return new
 
     @staticmethod
-    def convert_to_tf_coords(array):
-        """From [x1, y1, width, height] to [y1, x1, y2, x2], where x1 is x_min, x2 is x_max
-
-        This is an extraneous step as the model outputs [y1, x1, y2, x2] but were converted to the API
-        output format - only to keep the interface of the sync API.
-        """
-        x1 = array[0]
-        y1 = array[1]
-        width = array[2]
-        height = array[3]
-        x2 = x1 + width
-        y2 = y1 + height
-        return [y1, x1, y2, x2]
-
-    @staticmethod
     def __load_model(model_path):
         """Loads a detection model (i.e., create a graph) from a .pb file.
